Remove old `InteractionBarEditorView`
